모임홈 반응형 적용 / QuickMenu 디자인 개선 / 전체모임 반응형 QA 반영 (#1027)
* feat: quickMenu 디자인 개선

* feat: 모임신청가이드 및 svg 되돌리기

* fix: search marginTop 변경

* chore: 번쩍 모임 -> 번쩍 (QuickMenu label 변경)

* feat: dead code 추적 설정 추가

* chore: 전체모임 반응형 QA 반영

* feat: 모임홈 반응형 적용

* chore: 빌드에러 해결
