+++ .eslintrc.json
@@ -6,7 +6,7 @@
     "node": true,
     "browser": true
   },
-  "plugins": ["@typescript-eslint", "prettier"],
+  "plugins": ["@typescript-eslint", "prettier", "import"],
   "extends": [
     "eslint:recommended",
     "plugin:prettier/recommended",
@@ -26,6 +26,7 @@
       {
         "endOfLine": "auto"
       }
-    ]
+    ],
+    "import/no-unused-modules": [1, { "unusedExports": true }]
   }
 }
